requirements.txt
# Core deps will be added as we go, but expect things like:
# ollama
# faster-whisper
# pyaudio or sounddevice
# piper-tts or coqui-tts
# pyqt6 or pyside6
# psutil
# pyautogui
ollama
pyyaml
faster-whisper
sounddevice
soundfile
numpy
pyttsx3